+++ patchwork-networking/src/main/java/com/patchworkmc/mixin/networking/accessor/LoginQueryRequestS2CPacketAccessor.java
@@ -1,0 +1,15 @@
+package com.patchworkmc.mixin.networking.accessor;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+import net.minecraft.client.network.packet.LoginQueryRequestS2CPacket;
+
+@Mixin(LoginQueryRequestS2CPacket.class)
+public interface LoginQueryRequestS2CPacketAccessor {
+	@SuppressWarnings("PublicStaticMixinMember")
+	@Invoker("<init>")
+	public static LoginQueryRequestS2CPacket create() {
+		throw new AssertionError("Mixin not applied");
+	}
+}
