+++ patchwork-networking/src/main/resources/patchwork-networking.accessor.mixins.json
@@ -4,6 +4,10 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "ClientConnectionAccessor",
+    "CustomPayloadC2SPacketAccessor",
+    "CustomPayloadS2CPacketAccessor",
+    "LoginQueryRequestS2CPacketAccessor",
+    "LoginQueryResponseC2SPacketAccessor",
     "ThreadExecutorAccessor"
   ],
   "injectors": {
